# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/token/createGuestConnectToken": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "取得測試令牌1.0",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "會員帳號",
                        "name": "account",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "幣別 ",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "遊戲代碼",
                        "name": "gameID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/token/delConnectToken": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "令牌登出1.5",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DelConnectTokenBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/token/getConnectTokenAmount": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "取得令牌餘額1.4",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "連線令牌",
                        "name": "connectToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/token/getConnectTokenInfo": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "取得令牌資訊1.3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "連線令牌",
                        "name": "connectToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/token/updateConnectTokenLocation": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "變更令牌位置1.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateTokenLocation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/transaction/rollIn": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "遊戲錢包轉至個人錢包5.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RollInHistory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/transaction/rollOut": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "個人錢包轉至遊戲錢包5.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RollHistory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1.0/connectToken/authorization": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "令牌登入1.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AuthConnectToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.AuthConnectToken": {
            "type": "object",
            "properties": {
                "clientIpAddress": {
                    "type": "string",
                    "minLength": 1
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "entity.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "$ref": "#/definitions/entity.HttpResponseStatus"
                }
            }
        },
        "entity.DelConnectTokenBody": {
            "type": "object",
            "properties": {
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "entity.HttpResponseStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "traceCode": {
                    "type": "string"
                }
            }
        },
        "entity.RollHistory": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "gameSequenceNumber": {
                    "type": "string",
                    "minLength": 1
                },
                "rollTime": {
                    "type": "string"
                },
                "takeAll": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "transID": {
                    "type": "string"
                }
            }
        },
        "entity.RollInHistory": {
            "type": "object",
            "properties": {
                "betMode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ]
                },
                "betTime": {
                    "type": "string"
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "currencyKindBet": {
                    "type": "number"
                },
                "currencyKindContribution": {
                    "type": "number"
                },
                "currencyKindJackPot": {
                    "type": "number"
                },
                "currencyKindPayout": {
                    "type": "number"
                },
                "currencyKindWinLose": {
                    "type": "number"
                },
                "freeGame": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "gameRoom": {
                    "type": "integer",
                    "minimum": 1
                },
                "gameSequenceNumber": {
                    "type": "string",
                    "minLength": 1
                },
                "sequenceID": {
                    "type": "string",
                    "minLength": 1
                },
                "serverTime": {
                    "type": "string"
                },
                "transID": {
                    "type": "string"
                },
                "turnTimes": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "entity.UpdateTokenLocation": {
            "type": "object",
            "properties": {
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "location": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        }
    }
}